+++ src/views/LoanDetails/LoanDetails.js
@@ -32,11 +32,13 @@ import {
 import { IfPermission, stripesConnect } from '@folio/stripes/core';
 import { effectiveCallNumber } from '@folio/stripes/util';
 
-import PatronBlockModal from '../../components/PatronBlock/PatronBlockModal';
+import PatronBlockModalWithOverrideModal from '../../components/PatronBlock/PatronBlockModalWithOverrideModal';
 import {
   getFullName,
   nav,
   getOpenRequestsPath,
+  getRenewalPatronBlocksFromPatronBlocks,
+  accountsMatchStatus,
 } from '../../components/util';
 import { itemStatuses, loanActions } from '../../constants';
 import {
@@ -54,7 +56,6 @@ class LoanDetails extends React.Component {
   static propTypes = {
     stripes: PropTypes.object.isRequired,
     resources: PropTypes.shape({
-      loanAccountsActions: PropTypes.object,
       loanActions: PropTypes.object,
       loanActionsWithUser: PropTypes.object,
     }).isRequired,
@@ -77,6 +78,7 @@ class LoanDetails extends React.Component {
     patronGroup: PropTypes.object,
     user: PropTypes.object,
     loanActionsWithUser: PropTypes.arrayOf(PropTypes.object),
+    loanAccountActions: PropTypes.arrayOf(PropTypes.object),
     loanPolicies: PropTypes.object,
     requestCounts: PropTypes.object,
     renew: PropTypes.func,
@@ -94,6 +96,7 @@ class LoanDetails extends React.Component {
 
   static defaultProps = {
     enableButton: () => {},
+    loanAccountActions: [],
   };
 
   constructor(props) {
@@ -144,26 +147,39 @@ class LoanDetails extends React.Component {
     this.setState({ nonRenewedLoansModalOpen: false });
   }
 
-  renew = async () => {
+  onRenew = async (additionalInfo = '') => {
     const {
       loan,
       user,
-      patronBlocks,
       renew,
-      mutator: { renewals },
+      mutator: {
+        renewals,
+      },
     } = this.props;
-    const countRenew = patronBlocks.filter(p => p.renewals || p.blockRenewals);
 
-    if (!isEmpty(countRenew)) return this.setState({ patronBlockedModal: true });
+    await renew([loan], user, additionalInfo);
 
-    await renew([loan], user);
     return renewals.replace({ ts: new Date().getTime() });
   }
 
+  renew = async () => {
+    const {
+      patronBlocks,
+    } = this.props;
+    const countRenew = getRenewalPatronBlocksFromPatronBlocks(patronBlocks);
+
+    if (!isEmpty(countRenew)) {
+      return this.setState({
+        patronBlockedModal: true,
+      });
+    }
+
+    return await this.onRenew();
+  }
+
   viewFeeFine() {
-    const { stripes, resources } = this.props;
-    const records = resources?.loanAccountsActions?.records ?? [];
-    const total = records.reduce((acc, { amount }) => (acc + parseFloat(amount)), 0);
+    const { stripes, loanAccountActions } = this.props;
+    const total = loanAccountActions.reduce((acc, { amount }) => (acc + parseFloat(amount)), 0);
 
     if (total === 0) return '-';
 
@@ -172,6 +188,7 @@ class LoanDetails extends React.Component {
     return stripes.hasPerm('ui-users.accounts')
       ? (
         <button
+          data-test-fee-fine-details-link
           className={css.feefineButton}
           onClick={(e) => this.feefinedetails(e)}
           type="button"
@@ -183,22 +200,32 @@ class LoanDetails extends React.Component {
   }
 
   feefinedetails = (e) => {
-    const { history, match: { params }, resources } = this.props;
-    const accounts = resources?.loanAccountsActions?.records ?? [];
+    const {
+      history,
+      match: { params },
+      loanAccountActions,
+    } = this.props;
+
+    if (loanAccountActions.length === 1) {
+      nav.onClickViewAccountActionsHistory(e, { id: loanAccountActions[0].id }, history, params);
+      return undefined;
+    }
+
     const loan = this.loan || {};
 
-    if (accounts.length === 1) {
-      nav.onClickViewAccountActionsHistory(e, { id: accounts[0].id }, history, params);
-    } else if (accounts.length > 1) {
-      const open = accounts.filter(a => a?.status?.name === 'Open') || [];
-      if (open.length === accounts.length) {
-        nav.onClickViewOpenAccounts(e, loan, history, params);
-      } else if (open.length === 0) {
-        nav.onClickViewClosedAccounts(e, loan, history, params);
-      } else {
-        nav.onClickViewAllAccounts(e, loan, history, params);
-      }
+    if (accountsMatchStatus(loanAccountActions, 'closed')) {
+      nav.onClickViewClosedAccounts(e, loan, history, params);
+    } else if (accountsMatchStatus(loanAccountActions, 'open')) {
+      nav.onClickViewOpenAccounts(e, loan, history, params);
+    } else {
+      nav.onClickViewAllAccounts(e, loan, history, params);
     }
+
+    return undefined;
+  };
+
+  onOpenPatronBlockedModal = () => {
+    this.setState({ patronBlockedModal: true });
   };
 
   onClosePatronBlockedModal = () => {
@@ -407,6 +434,7 @@ class LoanDetails extends React.Component {
         />
       </p>
     );
+    const patronBlocksForModal = getRenewalPatronBlocksFromPatronBlocks(patronBlocks);
 
     return (
       <div data-test-loan-actions-history>
@@ -612,7 +640,7 @@ class LoanDetails extends React.Component {
               <Col xs={2}>
                 <KeyValue
                   data-test-loan-fees-fines
-                  label={<FormattedMessage id="ui-users.loans.details.fine" />}
+                  label={<FormattedMessage id="ui-users.loans.details.fineIncurred" />}
                   value={this.viewFeeFine()}
                 />
               </Col>
@@ -668,10 +696,12 @@ class LoanDetails extends React.Component {
                   </li>))
             }
             </Modal>
-            <PatronBlockModal
-              open={patronBlockedModal}
-              onClose={this.onClosePatronBlockedModal}
-              patronBlocks={patronBlocks.filter(p => p.renewals || p.blockRenewals)}
+            <PatronBlockModalWithOverrideModal
+              patronBlockedModalOpen={patronBlockedModal}
+              onClosePatronBlockedModal={this.onClosePatronBlockedModal}
+              onOpenPatronBlockedModal={this.onOpenPatronBlockedModal}
+              onRenew={this.onRenew}
+              patronBlocks={patronBlocksForModal}
               viewUserPath={`/users/view/${(user || {}).id}?filters=pg.${patronGroup.group}&sort=name`}
             />
             { this.props.user && this.renderChangeDueDateDialog() }